更新readme 解决一些兼容问题，更好的支持archlinux和centos类发行版。 对没有托盘的桌面环境提供更好的支持。

+++ build.sh
@@ -10,23 +10,29 @@ DIR="$( cd -P "$( dirname "$SOURCE"  )" && pwd  )"
 
 cd $DIR
 rm -rf dist/mydict
-pyinstaller mydict.spec
 
-version=`./dist/mydict/mydict -v`
+which dpkg >/dev/null 2>&1
+if [ $? -eq 0 ]; then
+    pyinstaller mydict.spec
 
-rm -rf deb/opt/my_dict >/dev/null 2>&1
-mkdir -p deb/opt/my_dict >/dev/null
-cp ./LICENSE ./deb/opt/my_dict
-cp -rf dist/mydict/* deb/opt/my_dict/
-sed -i "s/^Version:.*/Version:${version}/g" deb/DEBIAN/control
-dpkg -b deb dist/mydict_linux_x64_${version}.deb
-rm -rf deb/opt/my_dict/*
-sed -i "s/^Version:.*/Version:0.0.0/g" deb/DEBIAN/control
+    version=`./dist/mydict/mydict -v`
+    rm -rf deb/opt/my_dict >/dev/null 2>&1
+    mkdir -p deb/opt/my_dict >/dev/null
+    cp ./LICENSE ./deb/opt/my_dict
+    cp -rf dist/mydict/* deb/opt/my_dict/
+    sed -i "s/^Version:.*/Version:${version}/g" deb/DEBIAN/control
+    dpkg -b deb dist/mydict_linux_x64_${version}.deb
+    rm -rf deb/opt/my_dict/*
+    sed -i "s/^Version:.*/Version:0.0.0/g" deb/DEBIAN/control
 
-cd $DIR
-pyinstaller -y 我的词典.spec
+else
+    pyinstaller -y 我的词典.spec
 
-cd dist
-version=`./我的词典 -v`
-cp ../LICENSE .
-tar -czf my_dict_linux_x64_${version}.tar.gz 我的词典 LICENSE
+    cd $DIR/dist
+    version=`./我的词典 -v`
+    cp ../LICENSE .
+    source /etc/os-release
+    tar -czf mydict_${ID}_linux_x64_${version}.tar.gz 我的词典 LICENSE
+
+fi
+
